refactor: deepen ai chat session module
